[AIRFLOW-6367] Move variable tests from test_core to models/test_variable
